% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MetaAnalyser.R
\name{MetaAnalyser}
\alias{MetaAnalyser}
\title{The Meta-Analyser}
\usage{
MetaAnalyser(dat, rstudio = FALSE)
}
\arguments{
\item{dat}{Meta-analysis data. This should be a data frame with three columns, called "name", "est" and "se" giving the study name, study-specific parameter estimates and corresponding standard errors respectively.  If omitted, the app is launched with the default \code{\link{magnesium}} data.}

\item{rstudio}{The default of FALSE opens the app in the system default web browser.}
}
\value{
None
}
\description{
An interactive application to visualise meta-analysis data as a physical weighing machine
}
\details{
Opens a web browser with the interactive application
}
\examples{
\dontrun{MetaAnalyser(magnesium)}
}
\references{
J. Bowden and C. Jackson "Weighing evidence with the Meta-Analyser"
}

