% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metasumm.R
\name{resd_default}
\alias{resd_default}
\title{Heterogeneity standard deviation in meta-analysis}
\usage{
resd_default(dat)
}
\arguments{
\item{dat}{Meta-analysis data. This should be a data frame with three columns, called "name", "est" and "se" giving the study name, study-specific parameter estimates and corresponding standard errors respectively.  If omitted, the app is launched with the default \code{\link{magnesium}} data.}
}
\value{
Estimated random effects standard deviation
}
\description{
Random effects standard deviation using the classic DerSimonian & Laird formula.
}
\examples{
resd_default(magnesium)
}

