% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/metasumm.R
\name{resd_dsl}
\alias{resd_dsl}
\title{Heterogeneity standard deviation in meta-analysis}
\usage{
resd_dsl(dat)
}
\arguments{
\item{dat}{Meta-analysis data. This should be a data frame with three columns, called "name", "est" and "se" giving the study name, study-specific parameter estimates and corresponding standard errors respectively.

Numeric or character study names are permitted.  If the data frame has more than three columns, the first three are used.  If the first three columns are called "name", "est" and "se" in some order, they are re-ordered appropriately, otherwise they are re-named.}
}
\value{
Estimated random effects standard deviation
}
\description{
Random effects standard deviation using the classic DerSimonian & Laird formula.
}
\examples{
resd_dsl(magnesium)
}

